+++ pkg/actions/nodegroup/drain.go
@@ -4,7 +4,10 @@ import (
 	"context"
 	"time"
 
+	"github.com/kris-nova/logger"
 	"github.com/weaveworks/eksctl/pkg/eks"
+	"golang.org/x/sync/errgroup"
+	"golang.org/x/sync/semaphore"
 
 	"github.com/weaveworks/eksctl/pkg/drain"
 )
@@ -21,14 +24,32 @@ type DrainInput struct {
 }
 
 func (m *Manager) Drain(ctx context.Context, input *DrainInput) error {
+	parallelLimit := int64(input.Parallel)
+	sem := semaphore.NewWeighted(parallelLimit)
+	logger.Info("starting parallel draining, max in-flight of %d", parallelLimit)
+
 	if input.Plan {
 		return nil
 	}
-	for _, n := range input.NodeGroups {
-		nodeGroupDrainer := drain.NewNodeGroupDrainer(m.clientSet, n, input.MaxGracePeriod, input.NodeDrainWaitPeriod, input.PodEvictionWaitPeriod, input.Undo, input.DisableEviction, input.Parallel)
-		if err := nodeGroupDrainer.Drain(ctx); err != nil {
-			return err
-		}
+
+	g, ctx := errgroup.WithContext(ctx)
+	for _, nodegroup := range input.NodeGroups {
+		nodegroup := nodegroup
+		g.Go(func() error {
+			nodeGroupDrainer := drain.NewNodeGroupDrainer(m.clientSet, nodegroup, input.MaxGracePeriod, input.NodeDrainWaitPeriod, input.PodEvictionWaitPeriod, input.Undo, input.DisableEviction, input.Parallel)
+			return nodeGroupDrainer.Drain(ctx, sem)
+		})
 	}
-	return nil
+	err := g.Wait()
+	if err != nil {
+		logger.Critical("Node group drain failed: %w", err)
+	}
+	waitForAllRoutinesToFinish(ctx, sem, parallelLimit)
+	return err
+}
+
+func waitForAllRoutinesToFinish(ctx context.Context, sem *semaphore.Weighted, size int64) {
+	if err := sem.Acquire(ctx, size); err != nil {
+		logger.Critical("failed to acquire semaphore while waiting for all routines to finish: %w", err)
+	}
 }
